requirements.txt
Pillow==8.2.0

markdown2~=2.4.0
tkhtmlview~=0.1.0.post1
numpy~=1.20.3
